Source code for Stefano, added Callbacks and renamed some files
