+++ services/authentication/databaseFile/users.json
@@ -1,0 +1,1 @@
+[{"username":"Michele","password":"PASSWORD1"},{"username":"Stefano","password":"PASSWORD2"},{"username":"Nicola","password":"PASSWORD3"},{"username":"Marco","password":"PASSWORD4"}]
