+++ services/authentication/src/main/java/com/wedriveu/services/authentication/entity/User.java
@@ -1,0 +1,28 @@
+package com.wedriveu.services.authentication.entity;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+/**
+ * Created by Michele on 12/07/2017.
+ */
+public class User {
+
+    private String username;
+    private String password;
+
+    public User(@JsonProperty("username") String username,
+                @JsonProperty("password") String password) {
+        this.username = username;
+        this.password = password;
+    }
+
+    public String getUsername() {
+        return this.username;
+    }
+
+    public String getPassword() {
+        return this.password;
+    }
+
+}
+
