+++ services/booking/databaseFile/bookings.json
@@ -1,0 +1,1 @@
+[{"bookingID":1,"date":61475365414000,"username":"Nicola","carLicencePlate":"MACCHINA1","destinationLatitude":20.0,"destinationLongitude":12.1,"sourceLatitude":30.0,"sourceLongitude":31.1,"bookingState":"started"},{"bookingID":2,"date":61472596953000,"username":"Marco","carLicencePlate":"MACCHINA2","destinationLatitude":19.9,"destinationLongitude":11.0,"sourceLatitude":29.9,"sourceLongitude":30.0,"bookingState":"processing"},{"bookingID":3,"date":61469914892000,"username":"Stefano","carLicencePlate":"MACCHINA3","destinationLatitude":18.8,"destinationLongitude":10.9,"sourceLatitude":28.8,"sourceLongitude":29.9,"bookingState":"complete"},{"bookingID":4,"date":61467142831000,"username":"Michele","carLicencePlate":"MACCHINA4","destinationLatitude":17.7,"destinationLongitude":9.8,"sourceLatitude":27.7,"sourceLongitude":28.8,"bookingState":"complete"}]
