--- services/booking/src/main/java/Main.java
@@ -1,13 +1,0 @@
-import entity.Booking;
-import entity.BookingStoreImpl;
-
-public class Main {
-
-    public static void main(String[] args) {
-        BookingStoreImpl obj = new BookingStoreImpl();
-        obj.mapEntityToJson();
-        Booking me;
-        me = obj.getBooking(1);
-    }
-
-}
