+++ services/booking/src/main/java/com/wedriveu/services/booking/Main.java
@@ -1,0 +1,15 @@
+package com.wedriveu.services.booking;
+
+import com.wedriveu.services.booking.entity.Booking;
+import com.wedriveu.services.booking.entity.BookingStoreImpl;
+
+public class Main {
+
+    public static void main(String[] args) {
+        BookingStoreImpl obj = new BookingStoreImpl();
+        obj.mapEntityToJson();
+        Booking me;
+        me = obj.getBooking(1);
+    }
+
+}
