--- services/shared/src/main/java/utilities/Util.java
@@ -1,16 +1,0 @@
-package utilities;
-
-/**
- * Created by Michele on 17/07/2017.
- */
-public class Util {
-
-    public final static String USERS_DATABASE_PATH = "users.json";
-    public final static String BOOKINGS_DATABASE_PATH = "D:\\bookings.json";
-    public final static String VEHICLES_DATABASE_PATH = "services/vehicle/databaseFile/vehicles.json";
-
-
-    public void log(String toLog){
-        System.out.println(toLog);
-    }
-}
