+++ services/vehicle/databaseFile/vehicles.json
@@ -1,0 +1,1 @@
+[{"carLicencePlate":"MACCHINA1","state":"broken","position":{"latitude":10.2,"longitude":13.2},"lastUpdate":61472777863000},{"carLicencePlate":"MACCHINA2","state":"available","position":{"latitude":11.2,"longitude":14.2},"lastUpdate":61470009792000},{"carLicencePlate":"MACCHINA3","state":"busy","position":{"latitude":15.2,"longitude":13.2},"lastUpdate":61467152206000},{"carLicencePlate":"MACCHINA4","state":"recharging","position":{"latitude":13.2,"longitude":16.2},"lastUpdate":61464382642000}]
