--- services/vehicle/src/main/java/boundary/CommunicationWithVehicles.java
@@ -1,10 +1,0 @@
-package boundary;
-
-import java.io.IOException;
-/**
- * Created by stefano.bernagozzi on 17/07/2017.
- */
-public interface CommunicationWithVehicles {
-    public double requestBatteryPercentage(String licensePlate) throws IOException;
-
-}
