+++ services/vehicle/src/main/java/com/wedriveu/services/vehicle/callback/Callback.java
@@ -1,0 +1,8 @@
+package com.wedriveu.services.vehicle.callback;
+
+/**
+ * Created by Michele on 18/07/2017.
+ */
+public interface Callback {
+    void onRequestBatteryPercentage(double percentage);
+}
