+++ vehicle/build.gradle
@@ -3,10 +3,13 @@ apply plugin: 'java'
 sourceCompatibility = 1.8
 
 repositories {
+    jcenter()
     mavenCentral()
 }
 
 dependencies {
     testCompile 'junit:junit:4.12'
     compile 'com.fasterxml.jackson.core:jackson-databind:2.8.9'
+    compile group: 'com.rabbitmq', name: 'amqp-client', version: '2.3.1'
+    compile project(path: ':services:shared')
 }
