+++ vehicle/src/main/java/com/wedriveu/vehicle/Callback.java
@@ -1,0 +1,7 @@
+package com.wedriveu.vehicle;
+
+/**
+ * Created by Michele on 18/07/2017.
+ */
+public interface Callback {
+}
